生成msprof.json的Ascend Hardware层以及连线的末端
# message auto-generated for no-merge-commit merge:
merge json_trace into master

【AR20240703476904】msprof.json交付件的Ascend Hardware层C化代码开发

Created-by: msprofDeveloper

Match-id-81163a0b4908d32cce0a90a3d4ef136e0f015667

+++ analysis/csrc/domain/data_process/ai_task/task_processor.h
@@ -21,7 +21,7 @@ namespace Domain {
 // start_time, duration, model_id, index_id, stream_id, task_id, context_id, batch_id, connection_id host_task_type,
 // device_task_type
 using OriAscendTaskData = std::vector<std::tuple<double, double, uint32_t, int32_t, uint32_t, uint32_t, uint32_t,
-                                                 uint32_t, uint32_t, std::string, std::string>>;
+                                                 uint32_t, uint64_t, std::string, std::string>>;
 class TaskProcessor : public DataProcessor {
 public:
     TaskProcessor() = default;
